+++ src/design/plone/contenttypes/restapi/deserializers/dxfields.py
@@ -1,25 +1,18 @@
 # -*- coding: utf-8 -*-
-from design.plone.contenttypes import _
-from design.plone.contenttypes.interfaces import IDesignPloneContenttypesLayer
-from plone.dexterity.interfaces import IDexterityContent
-from plone.formwidget.geolocation.geolocation import Geolocation
 from plone.formwidget.geolocation.interfaces import IGeolocationField
-from plone.restapi.deserializer.dxfields import DefaultFieldDeserializer
-from plone.restapi.interfaces import IBlockFieldDeserializationTransformer
+from plone.dexterity.interfaces import IDexterityContent
 from plone.restapi.interfaces import IFieldDeserializer
 from zope.component import adapter
-from zope.component import subscribers
-from zope.i18n import translate
 from zope.interface import implementer
-from zope.schema.interfaces import ISourceText
-
-import json
-
-KEYS_WITH_URL = ["linkUrl", "navigationRoot", "showMoreLink"]
+from zope.publisher.interfaces.browser import IBrowserRequest
+from plone.restapi.deserializer.dxfields import DefaultFieldDeserializer
+from plone.formwidget.geolocation.geolocation import Geolocation
+from design.plone.contenttypes import _
+from zope.i18n import translate
 
 
 @implementer(IFieldDeserializer)
-@adapter(IGeolocationField, IDexterityContent, IDesignPloneContenttypesLayer)
+@adapter(IGeolocationField, IDexterityContent, IBrowserRequest)
 class GeolocationFieldDeserializer(DefaultFieldDeserializer):
     def __call__(self, value):
         if "latitude" not in value or "longitude" not in value:
@@ -33,48 +26,4 @@ class GeolocationFieldDeserializer(DefaultFieldDeserializer):
                     context=self.request,
                 )
             )
-        return Geolocation(
-            latitude=value["latitude"], longitude=value["longitude"]
-        )
-
-
-@implementer(IFieldDeserializer)
-@adapter(ISourceText, IDexterityContent, IDesignPloneContenttypesLayer)
-class SourceTextDeserializer(DefaultFieldDeserializer):
-    def __call__(self, value):
-        value = super(SourceTextDeserializer, self).__call__(value)
-        if self.field.getName() == "search_sections":
-            #  per ora solo con questo, ma potenzialmente con altri simili
-            data = json.loads(value)
-            for root in data:
-                for tab in root.get("items", []):
-                    for key in KEYS_WITH_URL:
-                        value = tab.get(key, [])
-                        if value:
-                            tab[key] = [
-                                x.get("UID", "")
-                                for x in value
-                                if x.get("UID", "")
-                            ]
-                    blocks = tab.get("blocks", {})
-                    if blocks:
-                        for id, block_value in blocks.items():
-                            block_type = block_value.get("@type", "")
-                            handlers = []
-                            for h in subscribers(
-                                (self.context, self.request),
-                                IBlockFieldDeserializationTransformer,
-                            ):
-                                if (
-                                    h.block_type == block_type
-                                    or h.block_type is None  # noqa
-                                ):
-                                    handlers.append(h)
-                            for handler in sorted(
-                                handlers, key=lambda h: h.order
-                            ):
-                                block_value = handler(block_value)
-
-                            blocks[id] = block_value
-            value = json.dumps(value)
-        return value
+        return Geolocation(latitude=value["latitude"], longitude=value["longitude"])