+++ CHANGES.rst
@@ -7,6 +7,10 @@ Changelog
 
 - Add search_sections field in control panel.
   [cekk]
+- Can add Images into Cartella Modulistica (to be able to add image blocks in it).
+  [cekk]
+- Customizable tipologie_documento.
+  [cekk]
 
 
 1.0.8 (2021-02-19)